Randomize on list instead on Proposal
